refactor: rename api folder to api-client folder and updaterelated imports
