+++ frontend/src/api-client/fetch-completed-matches.ts
