+++ frontend/src/api-client/fetch-draft-prediction.ts
