+++ frontend/src/api-client/fetch-hero-data.ts
