+++ frontend/src/api-client/fetch-league-info.ts
