+++ frontend/src/api-client/fetch-prediction-details.ts
